Add files via upload
chatbox UI looks better and streamlined.

owners of a listing can delete, but its not showing visibly, so I have to fix this. 

if login session persistence is still active, then the user will be redirected to default landing page (dashboard.jsx) instead of being in the login screen with the option to logout via righthand of navbar.

navbar updated for user roles, however, I will not use them yet. maybe for admin.